🎨 style(permission_ui.dart): refactor UI layout to use Row and Column widgets The UI layout for the PermissionPage widget has been refactored to use Row and Column widgets. This improves the readability of the code and makes it easier to understand the layout of the UI.
🐛 fix(presence_maps_ui_button.dart): fix animation controller mixin
The PresenceMapButtonState class now uses TickerProviderStateMixin instead of SingleTickerProviderStateMixin. This is because the PresenceMapButtonState class now has two AnimationControllers.

✨ feat(presence_maps_ui_button.dart): add pulsing animation to button
A pulsing animation has been added to the PresenceMapButton widget. This animation is achieved by using an AnimationController and a ScaleTransition widget. The animation is triggered when the user presses and holds the button.

🎨 style(presence_maps_ui.dart): change CherryToast animation and position
The CherryToast widget now uses a different